+++ ci/release_notes.md
@@ -1,1 +1,3 @@
-Empty - please add release notes here
+## Improvements
+- Add `--ignore-queue` to apply the head version of the previous environment.
+- Many small wording improvements.